mqtt/update-json/requirements.txt
psycopg2==2.9.6
paho_mqtt==1.6.1
pyaml